ipa: raspberrypi: Switch ipa and cam_helper to use std::chrono::duration
Switch the ipa and cam_helper code to use std::chrono::duration for
all time based variables. This improves code readability and avoids
possible errors when converting between time bases.

A new Duration typedef is defined to represent a std::chrono::duration
type with double precision representation and a nanosecond timebase that
will be used throughout. This minimises the loss of precision when
converting timebases. Macros for returning and streaming the duration
count are also provided.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>