pipeline: raspberrypi: Account for a missing Unicam embedded data node
The unicam driver no longer regesters an embedded data node if the sensor does
not provide this stream. Account for this in the pipeline handler match routine
by not assuming it is always present.

Add a warning if Unicam and the CamHelper do not agree on the presense of sensor
embedded data, and disable its useage in these cases.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>