libcamera: media_device: Add enumerateMediaWalks() helper
Add a new helper function to the MediaDevice class which returns a vector of
all possible media device topologies starting at the sensor entity and walking
to an endpoint for the media device. Each of these topologies is called a
MediaWalk. For each entity in a MediaWalk, we store a MediaEntity pointer
together with the source and sink MediaLink pointers.

As an example, consider the media graph below:

+----------+
|   CSI2   |
 +-----^----+
       |
   +---+---+
   |  Mux1 <-------+
   +--^----+       |
      |            |
+-----+---+    +---+---+
| Sensor1 |    |  Mux2 |<--+
+---------+    +-^-----+   |
                 |         |
         +-------+-+   +---+-----+
         | Sensor2 |   | Sensor3 |
         +---------+   +---------+

This would return three MediaDevice::MediaWalk structures looking like:

1)
+----------+
|   CSI2   |
+-----^----+
      |
+-----+----+
|   Mux1   |
+-----^----+
      |
+-----+----+
| Sensor1  |
+----------+

2)
+----------+
|   CSI2   |
+-----^----+
      |
+-----+----+
|   Mux1   |
+-----^----+
      |
+-----+----+
|   Mux2   |
+-----^----+
      |
+-----+----+
| Sensor2  |
+----------+

3)
+----------+
|   CSI2   |
+-----^----+
      |
+-----+----+
|   Mux1   |
+-----^----+
      |
+-----+----+
|   Mux2   |
+-----^----+
      |
+-----+----+
| Sensor3  |
+----------+

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>