pipeline: ipa: raspberrypi: Use IPA cookies
Pass an IPA cookie from the IPA to the pipeline handler through the
setDelayedControls signal. This cookie is the index of the current
RPiController::Metadata context used for the frame. The cookie is subsequently
passed into DelayedControls together with the controls to action.

The IPA cookie is then returned from DelayedControls when the frame with the
applied controls has been returned from the sensor, and eventually passed back
to the IPA from the signalIspPrepare signal.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>