camera_sensor: Add minimum and maximum line length to IPACameraSensorInfo
Add fields for minimum and maximum line length (in units of pixels) to the
IPACameraSensorInfo structure. This replaces the existing lineLength field.

Update the ipu3, raspberrypi and rkisp1 IPAs to use IPACameraSensorInfo::minLineLength
instead of IPACameraSensorInfo::lineLength, as logically we will always want to
use the fastest sensor readout by default.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>