ipa: raspberrypi: alsc: Small and mostly cosmetic improvements
This commit merely changes the order of the sparse matrices back to
how they were originally. They're now XY high and 4 wide, rather than
4 high and XY across. There should be no functional change.

This makes the code look again like it originally did, means that when
we access values in the same row they're not scattered across memory
so much, and the table resizing becomes just a single loop.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>