ipa: raspberrypi: Replace setMaxShutter with setSensorLimits in AGC
Create a SensorLimits structure in the AgcAlgorithm class. The maximum
shutter time is stored in this structure from the IPA. In a subsequent
commit, all gain and shutter limits will additionally be stored.

Pass the SensorLimits structure with the maximum shutter time into the
AGC algorithm through the new setSensorLimits() member function,
replacing the call to setMaxShutter().

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: David Plowman <david.plowman@raspberrypi.com>