libcamera: controls: Add a control for IQ stability
The IqUnstable metadata can be used by IPAs to indicate to an
application that they have not settled sufficiently to produce
reliable image quality. Applications would be advised to avoid using
frames flagged in this way.

One example would be when the camera starts, when the AEC/AGC might
oscillate for a few frames.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>